requirements.txt
streamlit
pymysql
ibm-watson-machine-learning
python-dotenv
pandas
plotly